+++ lib/okuribito.rb
@@ -1,4 +1,5 @@
 require "okuribito/version"
+require "okuribito/patch_module"
 require "yaml"
 require "active_support"
 require "active_support/core_ext"
@@ -8,33 +9,6 @@ module Okuribito
     CLASS_METHOD_SYMBOL = ".".freeze
     INSTANCE_METHOD_SYMBOL = "#".freeze
     PATTERN = /\A(?<symbol>[#{CLASS_METHOD_SYMBOL}#{INSTANCE_METHOD_SYMBOL}])(?<method_name>.+)\z/
-
-    module SimplePatchModule
-      private
-
-      def define_patch(method_name, _patch, _id, _opt = {})
-        define_method(method_name) do |*args|
-          yield(to_s, caller) if block_given?
-          super(*args)
-        end
-      end
-    end
-
-    module FunctionalPatchModule
-      private
-
-      def define_patch(method_name, patch, id, opt = {})
-        sn = method_name.to_s.gsub(/\?/, "__q").gsub(/!/, "__e").gsub(/=/, "__eq")
-        patch.instance_variable_set("@#{sn}_#{id}_called", false)
-        define_method(method_name) do |*args|
-          if block_given? && !patch.instance_variable_get("@#{sn}_#{id}_called")
-            yield(to_s, caller)
-            patch.instance_variable_set("@#{sn}_#{id}_called", true) if opt[:once_detect]
-          end
-          super(*args)
-        end
-      end
-    end
 
     def initialize(opt = {}, &callback)
       @callback = callback
